restructure: sidebar 컴포넌트 widget 폴더로 이동
